Move & update import to App & App tests
